requirements.txt
Flask==1.0.2
line-bot-sdk==1.8.0
psycopg2==2.6.7
Flask-SQLAlchemy==2.3
Flask-Migrate==2.3.1
Flask-Script==2.0.6